@@ -10,12 +10,12 @@ final class TypeMapper
 	/**
 	 * @var (callable(string $typeName, string $location): (string|Type|null))[]
 	 */
-	private $matchers = [];
+	private array $matchers = [];
 
 	/**
 	 * @var (callable(mixed $value, string $typeName): mixed)[]
 	 */
-	private $mappings = [];
+	private array $mappings = [];
 
 
 	/**
@@ -24,12 +24,12 @@ final class TypeMapper
 	 *
 	 * TODO: isn't it too general? Or just map db-type name to php type and back?
 	 */
-	public function addMapping(callable $typeMatcher, callable $mapper) {
+	public function addMapping(callable $typeMatcher, callable $mapper): void {
 		$this->matchers[] = $typeMatcher;
 		$this->mappings[] = $mapper;
 	}
 
-	public function map(string $location, string $typeName, $value) {
+	public function map(string $location, string $typeName, mixed $value): mixed {
 		if ($value === NULL) {
 			return NULL; // todo: really do not translate?
 		}
@@ -44,14 +44,7 @@ final class TypeMapper
 		throw CouldNotMapTypeException::didYouRegisterTypeMapperFor($typeName, $value);
 	}
 
-	private function getTypeForValue($value): string {
-		return !\is_object($value) ? \gettype($value) : \get_class($value);
-	}
-
-	/**
-	 * @return string|Type
-	 */
-	public function mapType(string $location, string $typeName)
+	public function mapType(string $location, string $typeName): string|Type
 	{
 		foreach($this->matchers as $idx => $matcher) {
 			if ( ($translatingType = $matcher($typeName, $location)) !== NULL) {
